Fix unsigned-wraparound pointer arithmetic in proc_push/pop/peek
UBSan (-fsanitize=undefined, Apple clang -O2) flagged 3 runtime errors
on the bootstrap compile path, all the same pattern:

  ta_inline.h:88 proc_push: addition of unsigned offset to P overflowed to P-0x48
  ta_inline.h:92 proc_pop:  addition of unsigned offset to P overflowed to P-0x80
  ta_inline.h:98 proc_peek: addition of unsigned offset to P overflowed to P-0x50

Cause: stack slots were addressed as (p->mem + p->mem_size + p->sp * sizeof(Val)).
sp is negative (stack grows down from the buffer top), but sizeof(Val) promotes
the product to size_t, so the offset wraps the GEP around the address space.
This is formally out-of-bounds pointer arithmetic (C11 6.5.6p8), which lets the
optimizer assume non-wrapping addresses - miscompilation risk that varies with
clang version/optimizer state.

Fix: fold the offset into one signed index first, matching the collision-check
expressions already used in the same functions:

  (Val *)(p->mem + (p->mem_size + p->sp * (int)sizeof(Val)))

Verification:
- UBSan build: 3 runtime errors -> 0
- Bootstrap fixed point: two consecutive make bootstrap byte-identical, and
  lib/bootstrap.tabc unchanged vs committed (semantic no-op)
- make test: all categories 0 failures

+++ ta_inline.h
@@ -85,17 +85,23 @@ static inline void proc_push(Proc *p, Val v) {
         }
         v = gc_root_pop(p);
     }
-    *(Val *)(p->mem + p->mem_size + p->sp * sizeof(Val)) = v;
+    /* NOTE: the slot address is computed as mem + (int offset) where the
+     * offset stays within [0, mem_size) — the stack grows down from the
+     * top of the buffer, so sp is negative but mem_size + sp*sizeof(Val)
+     * is a non-negative in-buffer offset. Writing it as a single signed
+     * index avoids the unsigned-wraparound pointer arithmetic UBSan
+     * flags (sp * sizeof(Val) promotes to size_t, wrapping the GEP). */
+    *(Val *)(p->mem + (p->mem_size + p->sp * (int)sizeof(Val))) = v;
 }
 
 static inline Val proc_pop(Proc *p) {
-    Val v = *(Val *)(p->mem + p->mem_size + p->sp * sizeof(Val));
+    Val v = *(Val *)(p->mem + (p->mem_size + p->sp * (int)sizeof(Val)));
     p->sp++;
     return v;
 }
 
 static inline Val proc_peek(Proc *p, int offset) {
-    return *(Val *)(p->mem + p->mem_size + (p->sp + offset) * sizeof(Val));
+    return *(Val *)(p->mem + (p->mem_size + (p->sp + offset) * (int)sizeof(Val)));
 }
 
 /* ============================================================
